Bug Tracking System Using Java
A web application called Bug Tracking System was created for software businesses in the JAVA programming language. The project's primary goal is to control errors and faults that arise during the various stages of software development. After this project is completed, the staff will be able to update the system, update issue descriptions, and resolve issues from any place with internet access.